# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI-powered Strudel code generation and collaborative live coding platform\n\nFeatures:\n- AI-powered Strudel code generation from natural language\n- Real-time collaborative editing via WebSockets\n- OAuth authentication (Google, GitHub)\n- Anonymous session support\n- Save and share Strudel patterns",
        "title": "Algorave API",
        "contact": {
            "name": "API Support",
            "url": "https://codeberg.org/algorave/server"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0"
    },
    "host": "algorave.dev",
    "paths": {
...
        "/api/v1/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get user's notifications with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Max notifications to return (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Only return unread notifications",
                        "name": "unread",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_rest_notifications.ListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/read-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark all user's notifications as read",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Mark all notifications as read",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the count of unread notifications for the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Get unread notification count",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api_rest_notifications.UnreadCountResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/{id}/read": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mark a specific notification as read",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Mark notification as read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/codeberg_org_algorave_server_internal_errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api_rest_notifications.ListResponse": {
            "type": "object",
            "properties": {
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api_rest_notifications.NotificationResponse"
                    }
                },
                "unread_count": {
                    "type": "integer"
                }
            }
        },
        "api_rest_notifications.NotificationResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "read": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api_rest_notifications.UnreadCountResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "codeberg_org_algorave_server_internal_errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "optional details (sanitized in production)",
                    "type": "string"
                },
                "error": {
                    "description": "error code (e.g., \"unauthorized\", \"not_found\")",
                    "type": "string"
                },
                "message": {
                    "description": "user-friendly message",
                    "type": "string"
                },
                "request_id": {
                    "description": "echoed from request for correlation",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token for authenticated requests. Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}